feat#10/: userId 토스트 메세지 출력
